Initial Commit of KiCAD Files
This includes both designs for 2 module HAT boards. It doesn't have a
copy of the libraries used, and some components need 3D models created,
but this would still be usable for designing a new version as it is.

This design includes a 44-pin version of the KE02 MCU as the IO expander
and stand-alone controller on the connector board. See
http://atomosys.com for more details.